requirements_full.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.10.0
cryptography==45.0.5
decorator==5.2.1
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
# Editable install with no version control (esop-fastapi==0.1.0)
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mypy==1.17.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
requests==2.32.4
rich==14.1.0
rsa==4.9.1
ruff==0.12.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlmodel==0.0.24
stack-data==0.6.3
starlette==0.47.2
traitlets==5.14.3
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250516
types-python-jose==3.5.0.20250531
typing-inspection==0.4.1
typing_extensions==4.14.1
ulid-py==1.1.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.32.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
